create IAnalyticsLogger interface, refactor
