+++ lib/entities/analytics-events.ts
