+++ lib/providers/analytics/index.ts
@@ -1,0 +1,7 @@
+import { AnalyticsEvent } from '../../entities';
+
+export interface IAnalyticsLogger {
+  sendAnalyticsEvent(analyticsEvent: AnalyticsEvent): Promise<void>;
+}
+
+export * from './firehose';
